// src/app/api/public/lost/PublicLostPortalItem.ts
export interface PublicLostPortalItem {
  atcId?: string;
  lstId?: string;
  MANAGE_ID?: string;
  lstLctNm?: string;
  lstPlace?: string;
  LS_PLC?: string;
  lstFilePathImg?: string;
  lstPrdtNm?: string;
  LS_THNG_DTLS?: string;
  lstSbjt?: string;
  NOTI_CONT?: string;
  lstYmd?: string;
  LS_DE?: string;
  prdtClNm?: string;
  THNG_CLASS_NM?: string;
  rnum?: string;
}